Change / add node auth types and integrate into existing function signatures
