+++ packages/backend/bindings/node/index.ts
@@ -51,6 +51,7 @@ import {
   getLegacySeedChecksum as _getLegacySeedChecksum
 } from '../../../shared/lib/typings/wallet'
 import { ClientOptions } from '../../../shared/lib/typings/client'
+import { NodeAuth } from '../../../shared/lib/typings/node'
 
 const addon = require('../index.node')
 const mailbox = []
@@ -220,8 +221,8 @@ export const api = {
     ((__ids: CommunicationIds) => Promise<string>) {
     return (__ids: CommunicationIds) => _sendMigrationBundle(sendMessage, __ids, nodes, bundleHash, mwm)
   },
-  getNodeInfo: function (accountId: AccountIdentifier, url?: string): ((__ids: CommunicationIds) => Promise<string>) {
-    return (__ids: CommunicationIds) => _getNodeInfo(sendMessage, __ids, accountId, url)
+  getNodeInfo: function (accountId: AccountIdentifier, url?: string, auth?: NodeAuth): ((__ids: CommunicationIds) => Promise<string>) {
+    return (__ids: CommunicationIds) => _getNodeInfo(sendMessage, __ids, accountId, url, auth)
   },
 
   // Event emitters
