+++ packages/shared/lib/profile.ts
@@ -1,14 +1,14 @@
 import { AvailableExchangeRates } from 'shared/lib/currency'
 import { persistent } from 'shared/lib/helpers'
 import { generateRandomId } from 'shared/lib/utils'
-import { asyncRemoveStorage, destroyActor, getStoragePath, getWalletStoragePath } from 'shared/lib/wallet'
+import { destroyActor, getStoragePath, getWalletStoragePath } from 'shared/lib/wallet'
 import { derived, get, Readable, writable } from 'svelte/store'
 import type { ChartSelectors } from './chart'
 import { Electron } from './electron'
 import {
     HistoryDataProps
 } from './marketData'
-import type { Node } from './typings/client'
+import type { Node } from './typings/node'
 
 export interface MigratedTransaction {
     address: string;
