+++ packages/shared/lib/typings/wallet.ts
@@ -32,26 +32,6 @@ export interface Duration {
     nanos: number
 }
 
-export interface NodeInfo {
-    nodeinfo: {
-        name: string;
-        version: string;
-        isHealthy: boolean;
-        networkId: string;
-        minPoWScore: number;
-        bech32HRP: string;
-        latestMilestoneIndex: number;
-        latestMilestoneTimestamp: number;
-        confirmedMilestoneIndex: number;
-        pruningIndex: number;
-        features: string[];
-        messagesPerSecond: number;
-        referencedMessagesPerSecond: number;
-        referencedRate: number;
-    }
-    url: string
-}
-
 export function backup(bridge: Bridge, __ids: CommunicationIds, destinationPath: string, password: string) {
     return bridge({
         actorId: __ids.actorId,
